Rename class Region to Roi, move to new roi.js
